add fixedGraphics container for fixed graphics
