+++ src/config/containers.js
@@ -4,5 +4,6 @@ var circleSelection = new PIXI.Container()
 var elements = new PIXI.Container()
 var backgroundChart = new PIXI.Container()
 var guideNumbers = new PIXI.Container()
+var fixedGraphics = new PIXI.Container()
 
-export { circleSelection, backgroundChart, elements, guideNumbers }
+export { circleSelection, backgroundChart, elements, guideNumbers, fixedGraphics }
